Optimize FRMS_Enabled check
This was iterating over every single loaded type to try and find the
FMRS.FMRS_Util type on every scene load. This one method takes about
1.7s of time during scene switch in my setup.

This changes the search for the FMRSType to be cached in a readonly
static during class init. It then replaces the typeOperation with a
search for the relevant assembly followed by a GetType call. This
changes what was a 1.7s search to something that should be effectively
instant in comparison.